Create Extensions folder
